edit user information & libarary added

+++ NodeJs/myapp/routes/categories.js
@@ -2,6 +2,7 @@ const express = require("express")
 const categoriesModel = require('../models/categories')
 const categoriesRouter = express.Router();
 const categoriesController = require('../controllers/categories')
+const booksModel = require('../models/books');
 
 categoriesRouter.post("/", async (req, res) => {
 
@@ -45,9 +46,9 @@ categoriesRouter.post("/", async (req, res) => {
         res.setHeader('Access-Control-Allow-Origin', 'http://localhost:4200');
         const category = await categoriesModel.find({}).populate("books").exec()
         try {
-           
+
             console.log(category);
-           
+
             res.json(category)
         }
         catch (err) {
@@ -67,15 +68,19 @@ categoriesRouter.post("/", async (req, res) => {
     })
 
     .delete("/:id", async (request, response) => {
-        const { id } = request.params
         try {
-            const deletedcategory = await categoriesModel.deleteOne({ _id: id })
-            response.send("categoriesModel Deleted Correctly")
+            const { id } = request.params
+            const category = await categoriesModel.findById(id);
+            if (category.books.length < 1) {
+                const deletedcategory = await categoriesModel.deleteOne({ _id: id });
+                response.status(200).send("categoriesModel Deleted Correctly")
+            } else {
+                response.send("Can't delete categroy contains books");
+            }
         } catch (error) {
-            return console.log(error);
+            response.status(400).send(error.message);
         }
     })
-
     .patch("/:id", async (request, response) => {
         const { id } = request.params;
         const category = request.body
